backend/requirements.txt
fastapi
uvicorn
network_as_code
python-dotenv